requirements.txt
#requests==2.32.5
python-dotenv==1.2.1
#beautifulsoup4==4.14.3
cloudscraper25==2.7.0
m3u8==6.0.0